requirements.txt
fastapi~=0.115.6
fastapi-mail
fastapi-limiter
aioredis~=2.0.1
sqlalchemy~=2.0.37
pydantic~=2.10.5
pydantic[email]
passlib[bcrypt]~=1.7.4
jose~=1.0.0
uvicorn~=0.34.0
fastapi-limiter
psycopg2-binary~=2.9.10
cloudinary~=1.42.1

python-jose~=3.3.0
dnspython~=2.7.0
idna~=3.10
anyio~=4.7.0
h11~=0.14.0
pip~=24.3.1
redis~=5.3.0b4
typing_extensions~=4.12.2
sniffio~=1.3.1
click~=8.1.8
MarkupSafe~=3.0.2
certifi~=2024.12.14
starlette~=0.41.3
email_validator~=2.2.0
pydantic_core~=2.27.2
greenlet~=3.1.1
psycopg2~=2.9.10
annotated-types~=0.7.0
Jinja2~=3.1.5
async-timeout~=5.0.1
python-dotenv~=1.0.1
pydantic-settings~=2.7.1
python-multipart~=0.0.20

pytest~=8.3.4
httpcore~=1.0.7
httpx~=0.28.1
config~=0.5.1
fakeredis~=2.26.2
pytest-mock~=3.14.0
alembic~=1.14.1